Extract Msf::DBManager::Import::IP260::ASPL
MSP-11124
